fix(model): add numeroAfiliado to Empleado, use BigDecimal for monetary fields and LocalDate for dates
- Add missing numeroAfiliado (nro_afiliado_aporta) to Empleado per ERD
- Change Double to BigDecimal for precioActual, totalVenta, precioUnitarioHistorico, subtotal
- Change java.util.Date to java.time.LocalDate for Venta.fecha

+++ src/main/java/com/grupoy/tpbd2/model/DetalleVenta.java
@@ -1,5 +1,7 @@
 package com.grupoy.tpbd2.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,6 +13,6 @@ public class DetalleVenta {
 
     private Producto producto;
     private Integer cantidad;
-    private Double precioUnitarioHistorico;
-    private Double subtotal;
+    private BigDecimal precioUnitarioHistorico;
+    private BigDecimal subtotal;
 }
